reverted structure updates - too many merge conflicts
