+++ src/data_streams.py
@@ -6,11 +6,6 @@ import handling_cameras
 import handling_paths_files 
 
 class DataStream(ABC):
-    """_summary_
-
-    Args:
-        ABC (_type_): _description_
-    """
     def __init__(self)->None:
         self.current_image = None
         
@@ -57,6 +52,11 @@ class CameraStream(DataStream):
         self.cam_op = handling_cameras.CameraOperator()
 
     def open_data_stream(self)->bool:
+        """setups and opens camera stream.
+
+        Returns:
+            bool: True, if successful. False, otherwise.
+        """
         if not (self.cam_op.select_camera_device()):
             return False
         if not (self.cam_op.open_camera_stream()):
@@ -65,10 +65,20 @@ class CameraStream(DataStream):
         return True
     
     def update_data_stream(self)->bool:
+        """updates camera stream i.e. the current image.
+
+        Returns:
+            bool: True, if successful. False, otherwise.
+        """
         self.current_image = self.cam_op.get_image_camera()
         return True
     
     def close_data_stream(self)->bool:
+        """closes camera stream.
+
+        Returns:
+            bool: True, if successful. False, otherwise.
+        """
         if not (self.cam_op.close_camera_stream()):
             return False
         return True
@@ -76,7 +86,7 @@ class CameraStream(DataStream):
 
 
 
-class FolderStream(DataStream):
+class ImageStream(DataStream):
     
     def __init__(self)->None:
         super().__init__()
@@ -85,6 +95,11 @@ class FolderStream(DataStream):
         self.image_list = None
     
     def open_data_stream(self)->bool:
+        """setups and opens image stream.
+
+        Returns:
+            bool: True, if successful. False, otherwise.
+        """
         input_path = self.ph.get_path_abs_input()
         if not (self.ph.check_path_validity(input_path)):
             return False
@@ -98,19 +113,22 @@ class FolderStream(DataStream):
         return True
     
     def update_data_stream(self)->bool:
+        """updates image stream i.e. the current image.
+
+        Returns:
+            bool: True, if successful. False, otherwise.
+        """
         amount_images = len(self.image_list)
         self.current_image = self.image_list[self._id_image % amount_images]
         self._id_image += 1
         return True
 
     def close_data_stream(self)->bool:
+        """closes image stream.
+
+        Returns:
+            bool: True, if successful. False, otherwise.
+        """
         self._id_image = 0
         self.update_data_stream()
         return True
-        
-    
-
-    
-    
-
-
